refactor(tui): extract frame preserve/evict helpers into frame-preserve module (<350 LOC)
